GSVSource refactor (#2933)
* refactor

* rename

* time mixin

* rename

* polytope reader

* simpler polytope

* refactor datatypes

* test

* refactor2

* refactor3

* fix docstring

* faster timeutil

* update timeutils to pandas

* speed up lazy dask creation

* info file

* no dependence on intake

---------
